Modify ServerErrorJsonApiResponse to include exception status code
There are scenarios where the HTTP Response Status Code needs to be set
based on the Exception Status Code.

Example:

3rd party service failures will be handled in the application layer,
creating custom Application Exceptions with appropriate Exception status
code (most likely 502 or 503).

This is something that should be used as a specific HTTP Response Status
Code instead of just using the general and too broad (and in this
specific case incorrect) status 500.